Integrate report filters into dashboard
