+++ lancamentos.html
@@ -14,7 +14,6 @@
       <li class="nav-item"><a class="nav-link" href="index.html">Dashboard</a></li>
       <li class="nav-item"><a class="nav-link" href="orcamento.html">Orçamento</a></li>
       <li class="nav-item"><a class="nav-link active" href="lancamentos.html">Lançamentos</a></li>
-      <li class="nav-item"><a class="nav-link" href="relatorios.html">Relatórios</a></li>
     </ul>
 
     <div class="d-flex mb-3">
