+++ orcamento.html
@@ -14,7 +14,6 @@
       <li class="nav-item"><a class="nav-link" href="index.html">Dashboard</a></li>
       <li class="nav-item"><a class="nav-link active" href="orcamento.html">Orçamento</a></li>
       <li class="nav-item"><a class="nav-link" href="lancamentos.html">Lançamentos</a></li>
-      <li class="nav-item"><a class="nav-link" href="relatorios.html">Relatórios</a></li>
     </ul>
 
     <div class="p-4 bg-white rounded shadow-sm">
