--- relatorios.html
@@ -1,62 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-BR">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Relatórios</title>
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet">
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.10.5/font/bootstrap-icons.css" rel="stylesheet">
-  <link rel="stylesheet" href="style.css">
-</head>
-<body class="bg-light">
-  <div class="container-fluid py-3">
-    <ul class="nav nav-tabs mb-3">
-      <li class="nav-item"><a class="nav-link" href="index.html">Dashboard</a></li>
-      <li class="nav-item"><a class="nav-link" href="orcamento.html">Orçamento</a></li>
-      <li class="nav-item"><a class="nav-link" href="lancamentos.html">Lançamentos</a></li>
-      <li class="nav-item"><a class="nav-link active" href="relatorios.html">Relatórios</a></li>
-    </ul>
-
-    <div class="d-flex mb-3">
-      <select id="yearSelect" class="form-select w-auto me-2"></select>
-      <div id="monthButtons" class="btn-group flex-wrap me-2"></div>
-      <select id="typeFilter" class="form-select w-auto">
-        <option value="Todos">Todos</option>
-        <option value="Receita">Receitas</option>
-        <option value="Despesa">Despesas</option>
-        <option value="Investimento">Investimentos</option>
-      </select>
-    </div>
-
-    <div class="row g-3">
-      <div class="col-md-6">
-        <div class="card h-100">
-          <div class="card-body">
-            <h5 class="card-title">Receitas x Despesas x Investimentos</h5>
-            <canvas id="reportChart"></canvas>
-          </div>
-        </div>
-      </div>
-      <div class="col-md-6">
-        <div class="card h-100">
-          <div class="card-body">
-            <h5 class="card-title">Saldo Mensal</h5>
-            <canvas id="reportSaldoChart"></canvas>
-          </div>
-        </div>
-      </div>
-      <div class="col-md-6">
-        <div class="card h-100">
-          <div class="card-body">
-            <h5 class="card-title">Categorias</h5>
-            <canvas id="reportPieChart"></canvas>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-
-  <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
-  <script type="module" src="js/relatorios.js"></script>
-</body>
-</html>
